Ugly proof of concept of mini preview window by setting tiny font, large window dimensions, and blocking some of the auto-resizing methods and messages.
Still a lot of work to make this viable: embed the window in the main window, block resizing loops that can happen with the backend, set a limit on large files. This POC uses evalfunc.c because it has over 10k lines. Also have not tested it with MMTextView (only MMCoreTextView so far).